requirements.txt
pyside6
opencv-python
pillow
pyautogui